requirements.txt
fastapi
pika
uvicorn[standard]
pydantic[email]
python-dotenv